feat: construct mvc hierarchy
