--- src/main/java/org/younginhambak/backend/archive/DocumentStatus.java
@@ -1,5 +1,0 @@
-package org.younginhambak.backend.archive;
-
-public enum DocumentStatus {
-  ACTIVE, DISABLED, DELETED, ARCHIVED
-}
