--- src/main/java/org/younginhambak/backend/archive/DocumentTagId.java
@@ -1,16 +1,0 @@
-package org.younginhambak.backend.archive;
-
-import jakarta.persistence.Embeddable;
-import lombok.*;
-
-import java.io.Serializable;
-
-@Embeddable
-@Getter
-@EqualsAndHashCode
-@AllArgsConstructor
-public class DocumentTagId implements Serializable {
-
-  private Long documentId;
-  private Long tagId;
-}
