+++ src/main/java/org/younginhambak/backend/archive/entity/DocumentStatus.java
@@ -1,0 +1,5 @@
+package org.younginhambak.backend.archive.entity;
+
+public enum DocumentStatus {
+  ACTIVE, DISABLED, DELETED, ARCHIVED
+}
