+++ src/main/java/org/younginhambak/backend/archive/entity/DocumentTagId.java
@@ -1,0 +1,16 @@
+package org.younginhambak.backend.archive.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.*;
+
+import java.io.Serializable;
+
+@Embeddable
+@Getter
+@EqualsAndHashCode
+@AllArgsConstructor
+public class DocumentTagId implements Serializable {
+
+  private Long documentId;
+  private Long tagId;
+}
