+++ src/main/java/org/younginhambak/backend/file/DocumentFile.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.younginhambak.backend.archive.Document;
+import org.younginhambak.backend.archive.entity.Document;
 
 import java.time.LocalDateTime;
 import java.util.EnumSet;
